fe_object10: decomp func_800BE924 — FieldEntityD pool init
Memsets the entity buffer for D_80085391 entries (stride 0x1B4), then
for each entry pulls a packed (upper:9, lower:7) value from D_800DE4E0,
derives rangeLo/rangeHi/pc/etc, and registers the entity in D_80085230
at indices [D_80085388+D_800852F8, ...). pc is masked with 0x7FFF.

Match keys: (1) `e = buf` declared OUTSIDE the inner scope-block so gcc
puts entity in s1 + lui-hi-of-D_80085391 in s0 (matching target);
(2) Read-back trick: `upper = e->rangeLo` reloads rangeLo from memory
after writing it, which keeps both `e->rangeHi = lower; e->rangeHi =
upper+lower+1` writes (gcc would otherwise DSE the first) and pins the
pc lookup to the memory-read value; (3) unk1AC/AB/AA = 0x80 placed
AFTER `e->pc =` so gcc centers the loop offset base at entity+0x1AA.

Also changes D_800DE4E0 type from `u8 *` to `u16 *` since it's the
script-bytecode stream (read as packed u16s); updates func_800BE30C
and func_800BE7F4 to match. FieldEntityD struct fleshed out with
flags/groupRanges/rangeHi/stackPtr/unk188/unk18A/unk1AA-AC.

+++ include/field.h
@@ -583,12 +583,24 @@ typedef struct {
  *        each tick ends with a call to @c func_800B2BA0.
  */
 typedef struct {
-    /* 0x000 */ u8  pad000[0x174];
+    /* 0x000 */ u8  pad000[0x160];
+    /* 0x160 */ s32 flags;
+    /* 0x164 */ u16 groupRanges[8];
     /* 0x174 */ u8  scriptGroup;
     /* 0x175 */ u8  activeMask;
     /* 0x176 */ u16 pc;
     /* 0x178 */ u16 rangeLo;
-    /* 0x17A */ u8  pad17A[0x3A];
+    /* 0x17A */ u16 rangeHi;
+    /* 0x17C */ u8  pad17C[0x08];
+    /* 0x184 */ s8  stackPtr;
+    /* 0x185 */ u8  pad185[0x03];
+    /* 0x188 */ s16 unk188;
+    /* 0x18A */ u16 unk18A;
+    /* 0x18C */ u8  pad18C[0x1E];
+    /* 0x1AA */ u8  unk1AA;
+    /* 0x1AB */ u8  unk1AB;
+    /* 0x1AC */ u8  unk1AC;
+    /* 0x1AD */ u8  pad1AD[0x07];
 } FieldEntityD; /* 0x1B4 */
 
 /** @brief Self/anchor pointer used by @c func_8009A8E0 after the
@@ -864,7 +876,7 @@ extern u8 D_800DE8D8;       /**< Mirror of @c EventHeader.field2 from the active
 extern u8 D_800DE8D9;       /**< Mirror of @c EventHeader.field0. */
 extern u8 D_800DE8DA;       /**< Mirror of @c EventHeader.field1. */
 extern u8 D_800DE8C0;       /**< Mirror of @c EventHeader.field3. */
-extern u8 *D_800DE4E0;      /**< Pointer to script bytecode start (header base + 8). */
+extern u16 *D_800DE4E0;     /**< Pointer to script bytecode start (header base + 8); read as packed u16s. */
 extern u8 *D_800DE4E4;      /**< Pointer to first sub-table (header base + offset4). */
 extern u8 *D_800DE4E8;      /**< Pointer to second sub-table (header base + offset6). */
